fix(pro): split manifestLoader to prevent RSC server deps leaking into SSR bundle
The shared manifestLoader.ts imported buildServerRenderer (react-server
condition only), which got pulled into the SSR server bundle via
getReactServerComponent.server.ts → webpack compilation errors.

Split into manifestLoader.ts (client renderer, safe for SSR) and
manifestLoaderServer.ts (server renderer, RSC-bundle-only). Also fixes
CacheDemoPage.tsx lint issues.
