+++ packages/react-on-rails-pro/src/cache/manifestLoader.ts
@@ -13,14 +13,12 @@
  */
 
 import type { BundleManifest } from 'react-on-rails-rsc';
-import { buildServerRenderer } from 'react-on-rails-rsc/server.node';
 import { buildClientRenderer } from 'react-on-rails-rsc/client.node';
 import loadJsonFile from '../loadJsonFile.ts';
 
 let clientManifestFileName: string | undefined;
 let serverClientManifestFileName: string | undefined;
 
-let serverRendererPromise: Promise<ReturnType<typeof buildServerRenderer>> | undefined;
 let clientRendererPromise: Promise<ReturnType<typeof buildClientRenderer>> | undefined;
 
 export function setManifestFileNames(clientManifest: string, serverClientManifest: string): void {
@@ -28,22 +26,8 @@ export function setManifestFileNames(clientManifest: string, serverClientManifes
   serverClientManifestFileName = serverClientManifest;
 }
 
-export function getServerRenderer(): Promise<ReturnType<typeof buildServerRenderer>> {
-  if (!serverRendererPromise) {
-    if (!clientManifestFileName) {
-      throw new Error(
-        'Manifest file names not set. Ensure setManifestFileNames() is called before getServerRenderer(). ' +
-          'This is done automatically during the first RSC render request.',
-      );
-    }
-    serverRendererPromise = loadJsonFile<BundleManifest>(clientManifestFileName)
-      .then((reactClientManifest) => buildServerRenderer(reactClientManifest))
-      .catch((err: unknown) => {
-        serverRendererPromise = undefined;
-        throw err;
-      });
-  }
-  return serverRendererPromise;
+export function getClientManifestFileName(): string | undefined {
+  return clientManifestFileName;
 }
 
 export function getClientRenderer(): Promise<ReturnType<typeof buildClientRenderer>> {
