+++ react_on_rails_pro/spec/dummy/client/app/ror-auto-load-components/CacheDemoPage.tsx
@@ -1,39 +1,32 @@
 import React, { Suspense } from 'react';
+// eslint-disable-next-line camelcase -- matches Next.js API naming convention
 import { unstable_cache } from 'react-on-rails-pro/cache';
 
-// Cached: same result returned on subsequent renders (timestamp stays the same)
+// eslint-disable-next-line camelcase
 const getCachedTimestamp = unstable_cache(
-  async (label: string) => {
-    return (
-      <span data-testid={`cached-${label}`}>
-        {label}: {new Date().toISOString()}
-      </span>
-    );
-  },
+  (label: string) => (
+    <span data-testid={`cached-${label}`}>
+      {label}: {new Date().toISOString()}
+    </span>
+  ),
   { id: 'timestamp', tags: ['timestamps'] },
 );
 
-// Cached with TTL: result expires after 5 seconds
+// eslint-disable-next-line camelcase
 const getCachedWithTTL = unstable_cache(
-  async () => {
-    return (
-      <span data-testid="ttl-value">
-        TTL value: {new Date().toISOString()} (revalidates every 5s)
-      </span>
-    );
-  },
+  () => (
+    <span data-testid="ttl-value">TTL value: {new Date().toISOString()} (revalidates every 5s)</span>
+  ),
   { id: 'ttl-timestamp', revalidate: 5 },
 );
 
-// Cached with a different tag for selective invalidation
+// eslint-disable-next-line camelcase
 const getCachedProduct = unstable_cache(
-  async (productId: number) => {
-    return (
-      <span data-testid={`product-${productId}`}>
-        Product #{productId} rendered at {new Date().toISOString()}
-      </span>
-    );
-  },
+  (productId: number) => (
+    <span data-testid={`product-${productId}`}>
+      Product #{productId} rendered at {new Date().toISOString()}
+    </span>
+  ),
   { id: 'product', tags: ['products'] },
 );
 
@@ -66,6 +59,7 @@ const CachedSection = async () => {
 };
 
 const UncachedSection = async () => {
+  // eslint-disable-next-line no-promise-executor-return
   await new Promise((resolve) => setTimeout(resolve, 10));
   return (
     <div>
@@ -76,25 +70,22 @@ const UncachedSection = async () => {
   );
 };
 
-const CacheDemoPage = () => {
-  return (
-    <div>
-      <h1>unstable_cache Demo</h1>
-      <p>
-        This page demonstrates <code>unstable_cache</code> from React on Rails Pro.
-        Cached sections retain their timestamps across page refreshes. Uncached sections
-        produce fresh timestamps on every render.
-      </p>
-      <hr />
-      <Suspense fallback={<div>Loading cached content...</div>}>
-        <CachedSection />
-      </Suspense>
-      <hr />
-      <Suspense fallback={<div>Loading uncached content...</div>}>
-        <UncachedSection />
-      </Suspense>
-    </div>
-  );
-};
+const CacheDemoPage = () => (
+  <div>
+    <h1>unstable_cache Demo</h1>
+    <p>
+      This page demonstrates <code>unstable_cache</code> from React on Rails Pro. Cached sections retain
+      their timestamps across page refreshes. Uncached sections produce fresh timestamps on every render.
+    </p>
+    <hr />
+    <Suspense fallback={<div>Loading cached content...</div>}>
+      <CachedSection />
+    </Suspense>
+    <hr />
+    <Suspense fallback={<div>Loading uncached content...</div>}>
+      <UncachedSection />
+    </Suspense>
+  </div>
+);
 
 export default CacheDemoPage;
